requirements.txt
Flask==3.0.3
Flask_Cors==4.0.0
numpy==1.23.3
Pillow==11.0.0
Requests==2.32.3
